scenarios/simple_scraper/pyproject.toml
[tool.poetry]
name = "simple-scraper"
version = "0.0.1"
description = ""
authors = ["ks6088ts <ks6088ts@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
